moved api link in config, added prompt for deletion

+++ src/components/ListView.js
@@ -4,7 +4,6 @@ import EditComponent from './EditComponent';
 
 export default class ListView extends Component {
 
-
     createNewList = (e) => {
         e.preventDefault();
         let newListName = e.target.newListName.value.trim();
@@ -17,7 +16,8 @@ export default class ListView extends Component {
 
     deleteClickHandler = (singleListName) => {
         // console.log('deleting ListName', singleListName)
-        this.props.onDeleteList(singleListName);
+        if (window.confirm(`Delete List "${singleListName.listName}?"`))
+            this.props.onDeleteList(singleListName);
     }
 
     editClickHandler = (singleList) => {
@@ -34,27 +34,33 @@ export default class ListView extends Component {
 
 
         return <div className="listView">
-            {(!this.props.list || this.props.list.length === 0) ?
-                "New List will appear here"
-                :
+            <div>
                 <ul>
+                    <h2>CATEGORY</h2>
                     {
-                        this.props.list.map((singleList) => {
-                            return <div key={singleList.listName} >
-                                <span onClick={() => { this.props.onListSelect(singleList) }}>{singleList.listName}</span>
-                                <EditComponent
-                                    onEditClickHandler={() => this.editClickHandler(singleList)}
-                                    onDeleteClickHandler={() => this.deleteClickHandler(singleList)}
-                                />
-                            </div>
-                        })
+                        (this.props.list.length === 0) ?
+                            "New Categories will appear here"
+                            :
+                            this.props.list.map((singleList) => {
+                                return <div className="content" key={singleList.id} >
+                                    <div className="pointer" onClick={() => { this.props.onListSelect(singleList) }}>{singleList.listName}</div>
+                                    <EditComponent
+                                        onEditClickHandler={() => this.editClickHandler(singleList)}
+                                        onDeleteClickHandler={() => this.deleteClickHandler(singleList)}
+                                    />
+                                    {/* <div className="liner"></div> */}
+                                </div>
+                            })
+
                     }
+                    <form onSubmit={this.createNewList}>
+                        <input type='text' name="newListName" />
+                        <button type="submit">New Category</button>
+                    </form>
                 </ul>
-            }
-            <form onSubmit={this.createNewList}>
-                <input type='text' name="newListName" />
-                <button type="submit">AddNewList</button>
-            </form>
+
+            </div>
+
         </div>
     }
